modified order controller

+++ server/src/controllers/order.js
@@ -1,4 +1,5 @@
 const Order = require('../models/order');
+const { createOrder } = require('../utilities/razorpay');
 
 const getOrder = async (req, res) => {
     const { id } = req.params.id
@@ -11,44 +12,143 @@ const getOrder = async (req, res) => {
     }
 };
 
+
+//TODO: see what we get in req.body
+// req.body: type, shop, items, print
 const addOrder = async (req, res) => {
     const order = req.body;
+    var customer = req.customer;
+    const num = "ord_" + Date.now().toString();
+    var total = 0;
+    for (let i = 0; i < order.items.length; i++) {
+        const item = await Item.findById(order.items[i]);
+        if (!item) return res.status(404).json({ message: 'Item not found' });
+        if (item.shop !== order.shop)
+            return res.status(401).json({ message: 'This item does not belong to your shop' });
+        total += item.price;
+    }
 
-    const newOrder = new Order(order);
+    const newOrder = new Order({
+        num,
+        type: order.type,
+        shop: order.shop,
+        customer: customer.id,
+        total,
+        items: order.items,
+    });
 
     try {
         await newOrder.save();
-        res.status(201).json(newOrder);
+        res.status(201).json({ newOrder });
     } catch (error) {
         res.status(409).json({ message: error.message })
     }
 };
 
+
+
 const updateOrderCustomer = async (req, res) => {
     const { id } = req.params;
-    const order = req.body;
-    if (order.customer !== req.customer.id)
+    const { items } = req.body;
+    const orderFromDB = await Order.findById(id);
+
+    if (orderFromDB.customer !== req.customer.id)
         return res.status(401).send('This order does not belong to you');
 
     if (!mongoose.Types.ObjectId.isValid(id))
         return res.status(404).send('No order with that id');
 
-    const updatedOrder = await Order.findByIdAndUpdate(id, item, { new: true });
-    res.json(updatedOrder);
+    if (order.status === 'unplaced') {
+        orderFromDB.items = items;
+        await orderFromDB.save();
+        return res.json(updatedOrder);
+    }
+
+    res.status(401).send('You can update only unplaced orders');
+};
+
+const placeOrder = async (req, res) => {
+    const { id } = req.params;
+    const orderFromDB = await Order.findById(id);
+    if (orderFromDB.customer !== req.customer.id)
+        return res.status(401).send('This order does not belong to you');
+
+    if (!mongoose.Types.ObjectId.isValid(id))
+        return res.status(404).send('No order with that id');
+
+    if (orderFromDB.status === 'unplaced') {
+
+        const data = await createOrder(orderFromDB.total, 'INR', orderFromDB.num, 1);
+        if (!data.status) return res.status(409).json({ message: data.error.message });
+        orderFromDB.payment.razorpay_order_id = data.data.id;
+        orderFromDB.status = 'pending';
+        await orderFromDB.save();
+        return res.json(orderFromDB);
+    }
+
+    res.status(401).send('You can only place unplaced orders');
 };
 
 const updateOrderShop = async (req, res) => {
     const { id } = req.params;
-    const order = req.body;
-    if (order.shop !== req.shop.id)
+    const { status } = req.body;
+
+    const orderFromDB = await Order.findById(id);
+    if (orderFromDB.shop !== req.shop.id)
         return res.status(401).send('This order does not belong to your shop');
 
     if (!mongoose.Types.ObjectId.isValid(id))
         return res.status(404).send('No order with that id');
 
-    const updatedOrder = await Order.findByIdAndUpdate(id, item, { new: true });
-    res.json(updatedOrder);
+    if (orderFromDB.status === 'Unplaced')
+        return res.status(401).send('You can only update placed orders');
+    if (orderFromDB.status === 'Cancelled')
+        return res.status(401).send('You can only update uncancelled orders');
+    if (orderFromDB.status === 'Delivered')
+        return res.status(401).send('You can only update undelivered orders');
+    if (orderFromDB.status === 'Pending' && status === 'Rejected') {
+        orderFromDB.status = status;
+        await orderFromDB.save();
+        //TODO: refund money
+        return res.json(orderFromDB);
+    }
+    if (orderFromDB.status === 'Pending' && status === 'Accepted') {
+        orderFromDB.status = status;
+        // generate otp for order
+        const otp = Math.floor(100000 + Math.random() * 900000);
+        orderFromDB.otp = otp;
+        await orderFromDB.save();
+        // update due_amount of shop
+        req.shop.razorpay.due_amount += orderFromDB.total;
+        await req.shop.save();
+        return res.json(orderFromDB);
+    }
+    if (orderFromDB.status === 'Accepted' && status === 'Ready') {
+        orderFromDB.status = status;
+        await orderFromDB.save();
+        return res.json(orderFromDB);
+    }
+    res.status(401).send('You can only update orders in the following order: Pending -> Accepted -> Ready -> Delivered');
 };
+
+const deliverOrder = async (req, res) => {
+    const { id } = req.params;
+    const { otp } = req.body;
+    const orderFromDB = await Order.findById(id);
+    if (orderFromDB.shop !== req.shop.id)
+        return res.status(401).send('This order does not belong to your shop');
+
+    if (!mongoose.Types.ObjectId.isValid(id))
+        return res.status(404).send('No order with that id');
+
+    if (orderFromDB.status === 'Ready' && orderFromDB.otp === otp) {
+        orderFromDB.status = 'Delivered';
+        await orderFromDB.save();
+        res.json(orderFromDB);
+    }
+    res.status(401).send('Incorrect OTP');
+};
+
 
 const deleteOrder = async (req, res) => {
     const { id } = req.params;
@@ -60,11 +160,15 @@ const deleteOrder = async (req, res) => {
     if (order.customer !== req.customer.id)
         return res.status(401).send('This order does not belong to you');
 
-    await order.delete();
-    res.json({ message: 'Order deleted successfully' });
+    if (order.status === 'unplaced') {
+        await order.delete();
+        return res.json({ message: 'Order deleted successfully' });
+    }
+
+    res.status(401).send('You can only delete unplaced orders');
 };
 
-module.exports = { getOrder, updateOrderCustomer, updateOrderShop, addOrder, deleteOrder };
+module.exports = { getOrder, updateOrderCustomer, updateOrderShop, addOrder, deleteOrder, placeOrder, deliverOrder };
